fix(domain): align Employee.id type with SQL schema (Int instead of Long)

+++ app/employee/EmployeeModel.scala
@@ -3,7 +3,7 @@ package employee
 import java.time.LocalDateTime
 
 case class Employee(
-  id: Option[Long] = None,
+  id: Option[Int] = None,
   firstName: String,
   lastName: String,
   email: String,
